requirements.txt
apprise==1.2.0
requests>=2.31
feedparser>=6.0.10
Flask>=2.3
python-dateutil>=2.8
pytest>=7.0
pytest-mock
flake8
